pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "crawlboy"
version = "1.0.0"
description = "Crawl sitemap URLs with Crawl4AI and export as Markdown"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Akshara Hegde" },
]
keywords = ["sitemap", "crawler", "markdown", "crawl4ai", "web-scraping"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Text Processing :: Markup :: Markdown",
]
dependencies = [
    "crawl4ai>=0.5.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "questionary>=2.0",
    "defusedxml>=0.7.1",
]

[project.scripts]
crawlboy = "crawlboy.crawler:main"

[project.urls]
Homepage = "https://github.com/aksharahegde/crawlboy"
Repository = "https://github.com/aksharahegde/crawlboy"
Issues = "https://github.com/aksharahegde/crawlboy/issues"


[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "hypothesis>=6.0",
    "bandit>=1.7.9",
    "pip-audit>=2.7.0",
    "ruff>=0.5.0",
]


[tool.pytest.ini_options]
addopts = "--cov=crawlboy --cov-report=term-missing --cov-report=xml --cov-fail-under=19"
testpaths = ["tests"]
